Refactor animation timing from layer root to layer theme
