+++ src/constants/project-carousel.constants.ts
