+++ src/main.tsx
@@ -1,4 +1,4 @@
-import "@/styles/index.css";
+import "@/styles/global.css";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.tsx";
